refactor(web): migrate recommendations/spotify/wrapped APIs to ReScript (Phase 3 cont.)
Adds axios `params` support to the foundation, then migrates three more
live modules.

Axios.res:
- configP<'params> + getP/postP/putP: pass axios `params` through so it
  serializes them into the query string (and drops undefined keys) exactly
  as the old `{ params: {...} }` did — no more hand-built query strings.
- authHeadersIfToken: Authorization only when a token exists (some endpoints
  send no auth header rather than "Bearer null").

Modules:
- Recommendations.res: trackRecommendation/artistRecommendation/
  albumRecommendation records; each getter returns `.data.<key> ?? []`
  (getOr(fromNullable(...), [])), with the conditional auth header.
- Spotify.res: play/pause (PUT) / next/previous (POST) / seek (PUT + params);
  responses were never consumed (callers are unused mutations) so they
  return unit; body is the empty object, matching `axios.put(url, {}, ...)`.
- Wrapped.res: the full WrappedData tree (7 nested record types) — the
  endpoint was untyped before; now getWrapped returns a concrete wrappedData.

limit defaults use ReScript labeled args (~limit=50); genType renders them as
a required `number | undefined` positional, so the 3 recommendation call
sites now pass `undefined` (runtime defaults to 50). Consumers switch to the
matching .gen modules; type imports take `<lower> as <PascalCase>`.

Verified: rescript build, tsc -b, full vite build. (Axios interop itself was
runtime-verified in the prior browser pass.)
