+++ apps/web/src/hooks/useRecommendations.tsx
@@ -3,12 +3,12 @@ import {
   getAlbumRecommendations,
   getArtistRecommendations,
   getTrackRecommendations,
-} from "../api/recommendations";
+} from "../api/Recommendations.gen";
 
 export const useTrackRecommendationsQuery = (did: string | undefined) =>
   useQuery({
     queryKey: ["trackRecommendations", did],
-    queryFn: () => getTrackRecommendations(did!),
+    queryFn: () => getTrackRecommendations(did!, undefined),
     enabled: !!did,
     staleTime: 5 * 60 * 1000,
   });
@@ -16,7 +16,7 @@ export const useTrackRecommendationsQuery = (did: string | undefined) =>
 export const useArtistRecommendationsQuery = (did: string | undefined) =>
   useQuery({
     queryKey: ["artistRecommendations", did],
-    queryFn: () => getArtistRecommendations(did!),
+    queryFn: () => getArtistRecommendations(did!, undefined),
     enabled: !!did,
     staleTime: 5 * 60 * 1000,
   });
@@ -24,7 +24,7 @@ export const useArtistRecommendationsQuery = (did: string | undefined) =>
 export const useAlbumRecommendationsQuery = (did: string | undefined) =>
   useQuery({
     queryKey: ["albumRecommendations", did],
-    queryFn: () => getAlbumRecommendations(did!),
+    queryFn: () => getAlbumRecommendations(did!, undefined),
     enabled: !!did,
     staleTime: 5 * 60 * 1000,
   });
